# Conflicts:
#	src/main/java/com/br/dreamday/controller/CadastroItemFornecedorWindow.java
#	src/main/java/com/br/dreamday/controller/DetalheFornecedorWindow.java
#	src/main/java/com/br/dreamday/dao/DaoItemFornecedor.java
#	src/main/java/com/br/dreamday/dao/postgres/DaoPostgresqlItemFornecedor.java
#	src/main/java/com/br/dreamday/domain/ItemFornecedor.java

+++ src/main/java/com/br/dreamday/controller/DetalheFornecedorWindow.java
@@ -1,6 +1,7 @@
 package com.br.dreamday.controller;
 
 import com.br.dreamday.MainViewApplication;
+import com.br.dreamday.domain.Categoria;
 import com.br.dreamday.domain.Fornecedor;
 import com.br.dreamday.domain.ItemFornecedor;
 import com.br.dreamday.service.FornecedorService;
@@ -14,6 +15,7 @@ import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
+import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.HBox;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
@@ -22,6 +24,18 @@ import java.io.IOException;
 import java.util.Objects;
 
 public class DetalheFornecedorWindow {
+
+    @FXML
+    private AnchorPane anchor;
+
+    @FXML
+    private Button btnEditar;
+
+    @FXML
+    private Button btnExcluir;
+
+    @FXML
+    private Button btnVincularItem;
 
     @FXML
     private TableView<ItemFornecedor> tableItens;
@@ -81,15 +95,26 @@ public class DetalheFornecedorWindow {
         popupStage.centerOnScreen();
         popupStage.setResizable(false);
         popupStage.showAndWait();
+
+        populaCampos(cadastroFornecedorWindow.getFornecedor());
     }
 
     @FXML
     void excluir(ActionEvent event) {
 
         confirmationMessage(() -> {
-            fornecedorService.excluirPor(fornecedor.getId());
-            Stage stage = (Stage) tableItens.getScene().getWindow();
-            stage.close();
+            try {
+                fornecedorService.excluirPor(fornecedor.getId());
+                Stage stage = (Stage) tableItens.getScene().getWindow();
+                stage.close();
+            } catch (Exception ex) {
+                exibirAlerta(
+                        Alert.AlertType.ERROR,
+                        "Erro ao deletar fornecedor",
+                        null,
+                        "Ocorreu um erro na exclusão do : " + ex.getMessage()
+                );
+            }
         });
     }
 
@@ -108,7 +133,7 @@ public class DetalheFornecedorWindow {
         FXMLLoader loader = new FXMLLoader(Objects.requireNonNull(MainViewApplication.class.getResource("cadastro-item-fornecedor-window.fxml")));
         Parent root = loader.load();
         CadastroItemFornecedorWindow cadastroItemFornecedorWindow = loader.getController();
-        cadastroItemFornecedorWindow.setAttributes(
+        cadastroItemFornecedorWindow.setAttributesInsercao(
                 new Fornecedor(fornecedor.getId(), fornecedor.getNome(), fornecedor.getTelefone(), fornecedor.getEmail())
         );
 
@@ -120,29 +145,24 @@ public class DetalheFornecedorWindow {
         popupStage.centerOnScreen();
         popupStage.setResizable(false);
         popupStage.showAndWait();
+        recarregarTabela();
+        tableItens.refresh();
     }
 
     private void mostrarTelaCadastroItemFornecedor(ItemFornecedor itemFornecedorSelecionado) throws IOException {
         FXMLLoader loader = new FXMLLoader(Objects.requireNonNull(MainViewApplication.class.getResource("cadastro-item-fornecedor-window.fxml")));
-        /*Parent root = loader.load();
-        DetalheFornecedorWindow detalheFornecedorWindow = loader.getController();
-        detalheFornecedorWindow.setAttributes(
-                new Fornecedor(
-                        fornecedorSelecionado.getId(),
-                        fornecedorSelecionado.getNome(),
-                        fornecedorSelecionado.getTelefone(),
-                        fornecedorSelecionado.getEmail()
-                )
-        );
+        Parent root = loader.load();
+        CadastroItemFornecedorWindow cadastroItemFornecedorWindow = loader.getController();
+        cadastroItemFornecedorWindow.setAttributesAlteracao(itemFornecedorSelecionado);
 
         Stage popupStage = new Stage();
-        popupStage.setTitle("Detalhe Fornecedor");
+        popupStage.setTitle("Alterar Item Fornecedor");
         Scene scene = new Scene(root);
         popupStage.setScene(scene);
         popupStage.initModality(Modality.APPLICATION_MODAL);
         popupStage.centerOnScreen();
         popupStage.setResizable(false);
-        popupStage.showAndWait();*/
+        popupStage.showAndWait();
     }
 
     public void exibirAlerta(Alert.AlertType tipo, String titulo, String cabecalho, String conteudo) {
@@ -158,22 +178,30 @@ public class DetalheFornecedorWindow {
         populaCampos(fornecedorSelecionado);
 
         itemFornecedorList = FXCollections.observableArrayList(itemFornecedorService.listarPor(fornecedor.getId()));
+        configuraColunasTabela();
+        configuraColunaAcoes();
+        tableItens.setItems(itemFornecedorList);
+    }
 
-        codigoColumn.setCellValueFactory(new PropertyValueFactory<>("id"));
-        descricaoColumn.setCellValueFactory(new PropertyValueFactory<>("produto.nome"));
+    private void configuraColunasTabela() {
+        codigoColumn.setCellValueFactory(new PropertyValueFactory<>("codigo"));
+        descricaoColumn.setCellValueFactory(new PropertyValueFactory<>("descricaoProduto"));
         precoColumn.setCellValueFactory(new PropertyValueFactory<>("preco"));
-        categoriaColumn.setCellValueFactory(new PropertyValueFactory<>("categoria.nome"));
+        categoriaColumn.setCellValueFactory(new PropertyValueFactory<>("nomeCategoria"));
+    }
 
+    private void configuraColunaAcoes() {
         acoesColumn.setCellFactory(column -> new TableCell<>() {
-            final Button editarButton = new Button("Detalhes");
+            final Button editarButton = new Button("Editar");
             final Button excluirButton = new Button("Excluir");
             final HBox buttonBox = new HBox(editarButton, excluirButton);
 
             {
-                buttonBox.setSpacing(10);
+                buttonBox.setSpacing(25);
 
                 editarButton.setOnAction(event -> {
                     ItemFornecedor itemFornecedor = getTableView().getItems().get(getIndex());
+                    itemFornecedor.setFornecedor(fornecedor);
 
                     try {
                         mostrarTelaCadastroItemFornecedor(itemFornecedor);
@@ -187,12 +215,21 @@ public class DetalheFornecedorWindow {
                     }
                 });
 
-                excluirButton.setOnAction(event -> {
-                    ItemFornecedor itemFornecedor = getTableView().getItems().get(getIndex());
-                    itemFornecedorService.excluirPor(itemFornecedor.getId());
-                    itemFornecedorList.remove(itemFornecedor);
-                    tableItens.refresh();
-                });
+                excluirButton.setOnAction(event -> confirmationMessage(() -> {
+                    try {
+                        ItemFornecedor itemFornecedor = getTableView().getItems().get(getIndex());
+                        itemFornecedorService.excluirPor(itemFornecedor.getId());
+                        itemFornecedorList.remove(itemFornecedor);
+                        tableItens.refresh();
+                    } catch(Exception ex) {
+                        exibirAlerta(
+                                Alert.AlertType.ERROR,
+                                "Exclusão de Item Fornecedor",
+                                null,
+                                "Ocorreu um erro ao deletar o item: " + ex.getMessage()
+                        );
+                    }
+                }));
             }
 
             @Override
@@ -205,8 +242,6 @@ public class DetalheFornecedorWindow {
                 }
             }
         });
-
-        tableItens.setItems(itemFornecedorList);
     }
 
     private void populaCampos(Fornecedor fornecedorSelecionado) {
@@ -227,5 +262,10 @@ public class DetalheFornecedorWindow {
                 action.run();
             }
         });
+    }
+
+    public void recarregarTabela() {
+        itemFornecedorList.clear();
+        itemFornecedorList.addAll(itemFornecedorService.listarPor(fornecedor.getId()));
     }
 }
